m2009/alpha_09.m m2009/driver_ensembles.m m2009/eis_reg_driver.m m2009/eis_trend_09.m global_wmean_script.m

+++ global_wmean_script.m
@@ -17,8 +17,8 @@ longit=vlon;
 latit=vlat;
 %nlatit=size(latit,1);
 % compute the weights
-latitweight=cos(pi/180*latit);
-glblatweight=cos(pi/180*latit);
+latitweight=cos(double(pi)/180*latit);
+glblatweight=cos(double(pi)/180*latit);
 %latitweight=1.*latit;
 %glblatweight=1.*latit;
 nlongit=length(longit);
